update wordle remove tic_tac_toe
