--- modules/tic_tac_toe/locales/zh_cn.json
@@ -1,11 +1,0 @@
-{
-    "tic_tac_toe.help.desc": "下井字棋。",
-    "tic_tac_toe.help": "和机器人下井字棋。",
-    "tic_tac_toe.help.expert": "和专家级机器人下井字棋。",
-    "tic_tac_toe.help.master": "和大师级机器人下井字棋。",
-    "tic_tac_toe.help.noob": "和菜鸟级机器人下井字棋。",
-    "tic_tac_toe.help.duo": "双人对战井字棋。",
-    "tic_tac_toe.message.turn": "请 ${player} 落子。\n（发送数字以确定要落子的位置）",
-    "tic_tac_toe.message.draw": "游戏结束，平局。",
-    "tic_tac_toe.message.winner": "游戏结束，${winner} 赢了。"
-}
