--- modules/tic_tac_toe/locales/zh_tw.json
@@ -1,11 +1,0 @@
-{
-    "tic_tac_toe.help.desc": "下井字棋。",
-    "tic_tac_toe.help": "和機器人下井字棋。",
-    "tic_tac_toe.help.expert": "和專家級機器人下井字棋。",
-    "tic_tac_toe.help.master": "和大師級機器人下井字棋。",
-    "tic_tac_toe.help.noob": "和菜鳥級機器人下井字棋。",
-    "tic_tac_toe.help.duo": "雙人對戰井字棋。",
-    "tic_tac_toe.message.turn": "請 ${player} 落子。\n（輸入數字以確定要落子的位置）",
-    "tic_tac_toe.message.draw": "游戏已结束，平局。",
-    "tic_tac_toe.message.winner": "游戏已结束，${winner} 赢了。"
-}
