+++ modules/wordle/locales/zh_cn.json
@@ -1,18 +1,18 @@
 {
+    "config.comments.wordle_disable_image": "wordle 模块是否禁用图片。",
     "wordle.help": "开始新游戏。",
     "wordle.help.desc": "玩 Wordle。",
-    "wordle.help.hard": "使用困难模式。",
+    "wordle.help.option.hard": "使用困难模式。",
+    "wordle.help.option.trial": "使用挑战模式。",
     "wordle.help.theme": "设置是否使用深色主题。",
     "wordle.message.finish": "游戏结束，答案为：${answer}。\n查询词典：https://dictionary.cambridge.org/dictionary/english-chinese-simplified/${answer}",
     "wordle.message.finish.success": "祝贺！你在 ${attempt} 回合内完成了游戏。",
-    "wordle.message.hard": "提示：正在使用困难模式，在上一回合中被颜色标记的字母在接下来的回合中必须出现，且必须符合对应的要求。",
     "wordle.message.hard.not_matched": "输入的单词不符合要求。",
     "wordle.message.not_a_word": "这不是一个有效的单词。",
     "wordle.message.start": "游戏开始，你需要猜一个由 5 个字母组成的单词，你有 6 次机会。\n绿色表示字母和位置都正确，黄色代表字母正确但位置不正确，灰色代表单词中没有这个字母。",
+    "wordle.message.start.hard": "在上一回合中被颜色标记的字母在接下来的回合中必须出现，且必须符合对应的要求。",
+    "wordle.message.start.trial": "你正在使用挑战模式，成功完成游戏可获得花瓣。一天只有一次机会！",
     "wordle.message.stop": "已停止，答案为：${answer}。\n查询词典：https://dictionary.cambridge.org/dictionary/english/${answer}",
     "wordle.message.theme.disable": "已关闭深色主题。",
-    "wordle.message.theme.enable": "已开启深色主题。",
-    "wordle.message.multi.warning": "提示：正在多人游戏中，无法获得花瓣。",
-    "wordle.option.multi": "多人游戏",
-    "config.comments.wordle_disable_image": "wordle 模块是否禁用图片。"
+    "wordle.message.theme.enable": "已开启深色主题。"
 }
