+++ modules/wordle/locales/zh_tw.json
@@ -5,7 +5,7 @@
     "wordle.help.theme": "設定是否使用深色主題。",
     "wordle.message.finish": "遊戲結束，答案為：${answer}。\n查詢辭典：https://dictionary.cambridge.org/dictionary/english-chinese-traditional/${answer}",
     "wordle.message.finish.success": "祝賀！你在 ${attempt} 回合內完成了遊戲。",
-    "wordle.message.hard": "提示：正在使用困難模式，在上一回合中被色彩標記的字母在接下來的回合中必須出現，且必須符合對應的要求。",
+    "wordle.message.hard": "在上一回合中被色彩標記的字母在接下來的回合中必須出現，且必須符合對應的要求。",
     "wordle.message.hard.not_matched": "輸入的單詞不符合要求。",
     "wordle.message.not_a_word": "這不是一個有效的單詞。",
     "wordle.message.start": "遊戲開始，你需要猜一個由 5 個字母組成的單詞，你有 6 次機會。\n綠色表示字母和位置都正確，黃色代表字母正確但位置不正確，灰色代表單詞中沒有這個字母。",
